+++ changelog.md
@@ -1,5 +1,11 @@
 ## Change log
 
+### 2.2.1
+* Include logout method.
+* Include set_server.
+* Return response text on add infrastructure with or without metadata.
+* Do nothing on set server if server is None.
+
 ### 2.2.0
 * Include get image observation by file, path, bytes or base64.